+++ src/plugin-hyperfy/utils.ts
@@ -1,7 +1,50 @@
+import { Readable } from 'node:stream';
+
 export async function hashFileBuffer(buffer: Buffer): Promise<string> {
     const hashBuf = await crypto.subtle.digest('SHA-256', buffer)
     const hash = Array.from(new Uint8Array(hashBuf))
       .map(b => b.toString(16).padStart(2, '0'))
       .join('')
     return hash
+}
+
+export async function convertToAudioBuffer(speechResponse: any): Promise<Buffer> {
+  if (Buffer.isBuffer(speechResponse)) {
+    return speechResponse;
+  }
+
+  if (typeof speechResponse?.getReader === 'function') {
+    // Handle Web ReadableStream
+    const reader = (speechResponse as ReadableStream<Uint8Array>).getReader();
+    const chunks: Uint8Array[] = [];
+
+    try {
+      while (true) {
+        const { done, value } = await reader.read();
+        if (done) break;
+        if (value) chunks.push(value);
+      }
+      return Buffer.concat(chunks);
+    } finally {
+      reader.releaseLock();
+    }
+  }
+
+  if (
+    speechResponse instanceof Readable ||
+    (speechResponse &&
+      speechResponse.readable === true &&
+      typeof speechResponse.pipe === 'function' &&
+      typeof speechResponse.on === 'function')
+  ) {
+    // Handle Node Readable Stream
+    return new Promise<Buffer>((resolve, reject) => {
+      const chunks: Buffer[] = [];
+      speechResponse.on('data', (chunk) => chunks.push(Buffer.from(chunk)));
+      speechResponse.on('end', () => resolve(Buffer.concat(chunks)));
+      speechResponse.on('error', (err) => reject(err));
+    });
+  }
+
+  throw new Error('Unexpected response type from TEXT_TO_SPEECH model');
 }
